Fix: Move integration test to standard test directory structure
- Moved CockroachDBConnectorIT.java from src/it/java to src/test/java
- Removed src/it directory structure to align with Debezium conventions
- Removed build-helper-maven-plugin configuration for src/it/java
- Simplified integration test to focus on basic connectivity and CRUD operations
- Removed EmbeddedEngine dependency to avoid compilation issues
- All unit tests continue to pass
